Blackjack partitions in Zig.

+++ blackjack/zig/outcomes.zig
@@ -1,14 +1,14 @@
 const std = @import("std");
 const io = std.io;
 
-fn partitions(cards: &[10]i8, subtotal: usize) usize {
-  var m: usize = 0;
+fn partitions(cards: &[10]i8, subtotal: u8) u32 {
+  var m: u32 = 0;
 
   // Hit
-  var i: usize = 0;
+  var i: u8 = 0;
   while (i < 10) : (i += 1) {
     if ((*cards)[i]>0) {
-      var total: usize = subtotal+i+1;
+      var total: u8 = subtotal+i+1;
       if (total < 21) {
         // Stand
         m += 1;
@@ -32,14 +32,14 @@ pub fn main() !void
   const stdout = &stdout_file_stream.stream;
 
   var deck: [10]i8 = []i8{4,4,4,4,4,4,4,4,4,16};
-  var d: usize = 0;
+  var d: u32 = 0;
   
-  var i: usize = 0;
+  var i: u8 = 0;
   while (i < 10) : (i += 1) {
     // Dealer showing
     deck[i] -= 1;
-    var p: usize = 0;
-    var j: usize = 0;
+    var p: u32 = 0;
+    var j: u8 = 0;
     while (j < 10) : (j += 1) {
       deck[j] -= 1;
       p += partitions(&deck, j+1);
